wiki/requirements.txt
Flask
cryptography
python-dotenv
requests
mysql-connector-python
